Updated WORKING VERSION

+++ src/com/JPA/onlineExam/repoTest/Book.java
@@ -12,7 +12,7 @@ public class Book {
 
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private int Id;
+	private Long Id;
 
 	private String bookName;
 
@@ -33,11 +33,11 @@ public class Book {
 		this.bookName = bookName;
 	}
 
-	public int getId() {
+	public Long getId() {
 		return Id;
 	}
 
-	public void setId(int id) {
+	public void setId(Long id) {
 		Id = id;
 	}
